fx_conservative_oanda/fx_conservative_oanda/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2
pytz>=2024.1
PyYAML>=6.0.1
oandapyV20>=0.6.3
pytest>=7.4.0
schedule>=1.2.0
